chore(cobra): migrate to cobra
migrates the command framework to cobra to support future subcommands
like `serve` to run a Github webhook to validate PRs
